+++ src/components/GridItem/GridItem.tsx
@@ -1,5 +1,5 @@
 import React, { useMemo } from 'react';
-import { getCommonProps } from '../../utils';
+import { BreakpointTokens, getCommonProps } from '../../utils';
 import classnames from 'classnames';
 import { determineColumnSpanClassName, validateColumnSpans } from './gridItemUtils';
 import { GridItemAlign } from './types';
@@ -10,12 +10,45 @@ export interface GridItemProps extends React.HTMLAttributes<HTMLDivElement> {
    */
   align?: GridItemAlign;
   /**
-   * column spans at different breakpoints, defaults to all columns.  If less than the total number of columns at the breakpoint it will be centered.
+   * Determines how many columns this GridItem spans at the xs breakpoint, defaults to the maximum of 2 columns.  If there are less than 2 columns in the Grid at the xs breakpoint it will be centered.
    */
   xs?: number;
+  /**
+   * Determines how many columns this GridItem spans at the sm breakpoint, defaults to the maximum of 2 columns.  If there are less than 2 columns in the Grid at the sm breakpoint it will be centered.
+   */
   sm?: number;
+  /**
+   * Determines how many columns this GridItem spans at the md breakpoint, defaults to the maximum of 6 columns.  If there are less than 6 columns in the Grid at the md breakpoint they will be centered.
+   */
   md?: number;
+  /**
+   * Determines how many columns this GridItem spans at the lg breakpoint, defaults to the maximum of 12 columns.  If there are less than 2 columns in the Grid at the lg breakpoint they will be centered.
+   */
   lg?: number;
+  /**
+   * Determines how many columns this GridItem spans at the xl breakpoint, defaults to the maximum of 12 columns.  If there are less than 2 columns in the Grid at the xl breakpoint they will be centered.
+   */
+  xl?: number;
+  /**
+   * The starting column for this GridItem at the xs breakpoint. If omitted, the GridItem will be placed in the next available column. Setting this value will override the alignment setting at the xs breakpoint.
+   */
+  xsColStart?: number;
+  /**
+   * The starting column for this GridItem at the sm breakpoint. If omitted, the GridItem will be placed in the next available column. Setting this value will override the alignment setting at the sm breakpoint.
+   */
+  smColStart?: number;
+  /**
+   * The starting column for this GridItem at the md breakpoint. If omitted, the GridItem will be placed in the next available column. Setting this value will override the alignment setting at the md breakpoint.
+   */
+  mdColStart?: number;
+  /**
+   * The starting column for this GridItem at the lg breakpoint. If omitted, the GridItem will be placed in the next available column. Setting this value will override the alignment setting at the lg breakpoint.
+   */
+  lgColStart?: number;
+  /**
+   * The starting column for this GridItem at the xl breakpoint. If omitted, the GridItem will be placed in the next available column. Setting this value will override the alignment setting at the xl breakpoint.
+   */
+  xlColStart?: number;
   /**
    * Optional element to render as the top-level component e.g. 'div', 'span', CustomComponent, etc.  Defaults to 'div'.
    */
@@ -37,6 +70,12 @@ const GridItem = ({
   sm = 2,
   md = 6,
   lg = 12,
+  xl = 12,
+  xsColStart,
+  smColStart,
+  mdColStart,
+  lgColStart,
+  xlColStart,
   align = GridItemAlign.center,
   element: Element = 'div',
   className,
@@ -44,16 +83,23 @@ const GridItem = ({
 }: GridItemProps) => {
   const { className: baseClassName, ...commonProps } = getCommonProps(props, 'GridItem');
 
-  const columnSpansPerBreakpoint = useMemo(() => ({ xs, sm, md, lg }) as const, [xs, sm, md, lg]);
+  const columnSpansPerBreakpoint = useMemo(() => ({ xs, sm, md, lg, xl }) as const, [xs, sm, md, lg, xl]);
+  const columnStartsPerBreakpoint = useMemo(
+    () => ({ xs: xsColStart, sm: smColStart, md: mdColStart, lg: lgColStart, xl: xlColStart }),
+    [xsColStart, smColStart, mdColStart, lgColStart, xlColStart],
+  );
+
   const gridItemClasses = useMemo(() => {
     return [
       baseClassName, // figure out the class names for each breakpoint
-      Object.entries(columnSpansPerBreakpoint).map(([key, value]) =>
-        determineColumnSpanClassName(key as GridItemAlign, value, align),
-      ),
+      Object.entries(columnSpansPerBreakpoint).map(([key, columnSpan]) => {
+        const columnStart = columnStartsPerBreakpoint[key as BreakpointTokens];
+
+        return determineColumnSpanClassName(key as BreakpointTokens, columnSpan, columnStart, align);
+      }),
       className,
     ];
-  }, [align, columnSpansPerBreakpoint, baseClassName, className]);
+  }, [baseClassName, columnSpansPerBreakpoint, className, columnStartsPerBreakpoint, align]);
 
   if (!validateColumnSpans(Object.values(columnSpansPerBreakpoint))) {
     return null;